Split conversion code over more files
