+++ plugins/formal-verification/src/org/jetbrains/kotlin/formver/conversion/ConvertedVar.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2010-2023 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlin.formver.conversion
+
+import org.jetbrains.kotlin.formver.scala.silicon.ast.*
+import viper.silver.ast.LocalVarDecl
+
+class ConvertedVar(val name: ConvertedName, val type: ConvertedType) {
+    fun toLocalVarDecl(
+        pos: Position = Position.NoPosition,
+        info: Info = Info.NoInfo,
+        trafos: Trafos = Trafos.NoTrafos,
+    ): LocalVarDecl = localVarDecl(name.asString, type.viperType, pos, info, trafos)
+
+    fun toLocalVar(
+        pos: Position = Position.NoPosition,
+        info: Info = Info.NoInfo,
+        trafos: Trafos = Trafos.NoTrafos,
+    ): Exp.LocalVar = Exp.LocalVar(name.asString, type.viperType, pos, info, trafos)
+
+    fun preconditions(): List<Exp> = type.preconditions(toLocalVar())
+    fun postconditions(): List<Exp> = type.postconditions(toLocalVar())
+}
